requirements.txt
sentence-transformers
pandas
scikit-learn
flask
numpy
python-dotenv
supabase
